fixes: gho bridge (#2029)
