+++ src/locales/en/messages.po
@@ -222,7 +222,7 @@ msgid "Allows you to switch between <0>variable</0> and <1>stable</1> interest r
 msgstr "Allows you to switch between <0>variable</0> and <1>stable</1> interest rates, where variable rate can increase and decrease depending on the amount of liquidity in the reserve, and stable rate will stay the same for the duration of your loan."
 
 #: src/components/transactions/AssetInput.tsx
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 #: src/components/transactions/Faucet/FaucetModalContent.tsx
 #: src/components/transactions/Switch/SwitchAssetInput.tsx
 #: src/modules/bridge/BridgeWrapper.tsx
@@ -243,7 +243,7 @@ msgstr "Amount in cooldown"
 msgid "Amount must be greater than 0"
 msgstr "Amount must be greater than 0"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "Amount to Bridge"
 msgstr "Amount to Bridge"
 
@@ -329,7 +329,7 @@ msgstr "Asset cannot be used as collateral."
 msgid "Asset category"
 msgstr "Asset category"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "Asset has been successfully sent to CCIP contract. You can check the status of the transactions below"
 msgstr "Asset has been successfully sent to CCIP contract. You can check the status of the transactions below"
 
@@ -592,7 +592,7 @@ msgstr "Borrowing {symbol}"
 msgid "Both"
 msgstr "Both"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 #: src/modules/bridge/BridgeWrapper.tsx
 msgid "Bridge Transactions"
 msgstr "Bridge Transactions"
@@ -605,7 +605,7 @@ msgstr "Bridge history"
 msgid "Bridge {symbol}"
 msgstr "Bridge {symbol}"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "Bridged Via CCIP"
 msgstr "Bridged Via CCIP"
 
@@ -1138,7 +1138,8 @@ msgstr "Faucet"
 msgid "Faucet {0}"
 msgstr "Faucet {0}"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "Fee"
 msgstr "Fee"
 
@@ -1189,7 +1190,7 @@ msgstr "French"
 msgid "Funds in the Safety Module"
 msgstr "Funds in the Safety Module"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "GHO balance"
 msgstr "GHO balance"
 
@@ -1796,7 +1797,7 @@ msgstr "Please always be aware of your <0>Health Factor (HF)</0> when partially 
 msgid "Please connect a wallet to view your personal information here."
 msgstr "Please connect a wallet to view your personal information here."
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "Please connect your wallet to be able to bridge your tokens."
 msgstr "Please connect your wallet to be able to bridge your tokens."
 
@@ -2968,7 +2969,7 @@ msgstr "Version 3"
 msgid "View"
 msgstr "View"
 
-#: src/components/transactions/Bridge/BridgeModal.tsx
+#: src/components/transactions/Bridge/BridgeModalContent.tsx
 msgid "View Bridge Transactions"
 msgstr "View Bridge Transactions"
