+++ src/services/WalletBalanceService.ts
@@ -1,9 +1,9 @@
 import { ChainId, WalletBalanceProvider } from '@aave/contract-helpers';
 import { normalize } from '@aave/math-utils';
 import { Provider } from '@ethersproject/providers';
+import { formatUnits } from 'ethers/lib/utils';
 import { governanceV3Config } from 'src/ui-config/governanceConfig';
 import { MarketDataType } from 'src/ui-config/marketsConfig';
-import { formatUnits } from 'ethers/lib/utils';
 
 export interface GovernanceTokensBalance {
   aave: string;
